+++ ts/packages/knowPro/src/conversationIndex.ts
@@ -20,10 +20,8 @@ import {
 import { IndexingEventHandlers } from "./interfaces.js";
 import { conversation as kpLib } from "knowledge-processor";
 import { openai } from "aiclient";
-import {
-    extractKnowledgeFromTextBatch,
-    facetValueToString,
-} from "./knowledge.js";
+import { extractKnowledgeFromTextBatch } from "./knowledge.js";
+import { facetValueToString } from "./common.js";
 import { createKnowledgeExtractor } from "./knowledge.js";
 import {
     addToSecondaryIndexes,
